revolico_publicador/requirements.txt
playwright==1.44.0
pandas==2.2.*
openpyxl==3.1.*
python-dotenv==1.0.*
requests==2.32.*
tqdm==4.66.*
colorama==0.4.*
gdown==5.*